bubble
Optimized bubble sort algorithm.  Space complexity is o(1) (constant time), because this algorithm is done inplace on the input array. No additional memory is used. Time complexity is o(n^2) in average/worst case because we are looping through the array multiple times until the array gets sorted. Time complexity in the best case is o(n), when the input is already sorted. We still need to traverse the array to confirm they are in the correct order, and ask ourselves 'did we perform any swaps? No we didn't, so we are complete'.